refactor(ci): split the gate probe's extractor from its classifier probe
Behaviour-neutral. Closes the headroom problem filed as GitHub issue #1803,
ahead of the round-10 findings, so the fix that follows is shaped by the
problem rather than by the line budget.

check-sentry-suites-in-ci-gate-extract.mjs takes the two lower layers: the
machinery for running a probe shell at all — fixed environment, probe-owned
directories, a deadline that reaches the whole process group, and an
interpreter resolved to an absolute path — and `bashFunctionSource`, which
asks bash where a function ends. What stays behind is the
classifier-specific part built on top: the guard, the read-time source
checks and `gateClassifications`. That direction avoids a cycle, since the
extractor needs the shells and the classifier probe needs both.

The move was made by partitioning the module into top-level blocks and
asserting the partition reassembles the original byte for byte before
anything was written, so no line could be dropped or reordered in the
process. Nothing else changed: 48 cases before, 48 cases after, on bash
3.2.57 and 5.3.15 both, with no test renamed.

  before  3.2 -> 48 pass, 5.3 -> 48 pass
  after   3.2 -> 48 pass, 5.3 -> 48 pass

Sizes go from 993/997 against the 1,000-line cap to 388/632 for the two
modules, which is the headroom the next change needs.

Refs #1779
Refs #1803
